Server/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
customtkinter==5.1.3
darkdetect==0.8.0
distlib==0.3.7
Django==4.2.4
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.12.2
pika==1.3.2
Pillow==9.5.0
pipenv==2023.11.15
platformdirs==3.10.0
protobuf==3.20.3
psycopg2-binary==2.9.9
pygame==2.5.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
setuptools==75.8.0
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.2
